+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Click to open query in mouse hover
-
+- Add `odata.openResultInNewPane` setting to control if query results open in a new pane (new default) or the active pane
 
 ## [0.3.0] - 2025-04-19